Add ability to create custom graphs with multiple metrics
Add ability to create custom graphs with multiple metrics:
--custom 'footcp:network.tcp.outrsts,network.tcp.ofoqueue'
The above would create a footcp  page with the two metrics combined in
a single graph. The general syntax should be:
--custom '<label>:<metric1>,<metric2>,...<metricN>'

Currently if you create a custom graph with two metrics that have the
same indoms it is impossible to distringuish which label (indom) is
associated with which metric. For example:
./pcpstats --include 'network.*' --graph 'foo:network.interface.in.bytes,network.interface.out.bytes' \
tests/pcp-files/server1.internal/20140510.08.47.0